+++ libs/obs/src/gnss_messages_common.cpp
@@ -127,54 +127,6 @@ gnss_message* gnss_message::readAndBuildFromStream(
 	return msg;
 }
 
-// Ctor (default: nullptr pointer)
-gnss_message_ptr::gnss_message_ptr() = default;
-// Ctor:Makes a copy of the pointee
-gnss_message_ptr::gnss_message_ptr(const gnss_message_ptr& o)
-{
-	if (!o.ptr) { ptr = nullptr; }
-	else
-	{
-		mrpt::io::CMemoryStream buf;
-		auto arch = mrpt::serialization::archiveFrom(buf);
-		o->writeToStream(arch);
-		buf.Seek(0);
-		ptr = gnss_message::readAndBuildFromStream(arch);
-	}
-}
-/** Assigns a pointer */
-gnss_message_ptr::gnss_message_ptr(const gnss_message* p)
-	: ptr(const_cast<gnss_message*>(p))
-{
-}
-void gnss_message_ptr::set(gnss_message* p)
-{
-	if (ptr)
-	{
-		delete ptr;
-		ptr = nullptr;
-	}
-	ptr = p;
-}
-// Makes a copy of the pointee
-gnss_message_ptr& gnss_message_ptr::operator=(const gnss_message_ptr& o)
-{
-	mrpt::io::CMemoryStream buf;
-	auto arch = mrpt::serialization::archiveFrom(buf);
-	o->writeToStream(arch);
-	buf.Seek(0);
-	ptr = gnss_message::readAndBuildFromStream(arch);
-	return *this;
-}
-gnss_message_ptr::~gnss_message_ptr()
-{
-	if (ptr)
-	{
-		delete ptr;
-		ptr = nullptr;
-	}
-}
-
 // ---------------------------------------
 UTC_time::UTC_time() = default;
 void UTC_time::writeToStream(mrpt::serialization::CArchive& out) const